+++ src/ask_for_help_bot/handler.py
@@ -26,8 +26,7 @@ from chatbot_core.v3.model.outgoing_event import OutgoingEvent, NotificationEven
 from common.button_payload import ButtonPayload
 from common.wenet_event_handler import WenetEventHandler
 from uhopper.utils.alert import AlertModule
-from wenet.common.interface.exceptions import TaskCreationError, RefreshTokenExpiredError, TaskTransactionCreationError, \
-    TaskNotFound
+from wenet.common.interface.exceptions import TaskCreationError, RefreshTokenExpiredError, TaskTransactionCreationError
 from wenet.common.model.message.event import WeNetAuthenticationEvent
 from wenet.common.model.message.message import TextualMessage, Message, QuestionToAnswerMessage, \
     AnsweredQuestionMessage, IncentiveMessage, IncentiveBadge, AnsweredPickedMessage
@@ -48,8 +47,11 @@ class AskForHelpHandler(WenetEventHandler):
     # context keys
     CONTEXT_DESIRED_ANSWERER = "desired_answerer"
     CONTEXT_DESIRED_ANSWERER_REASON = "desired_answerer_reason"
+    CONTEXT_SENSITIVE_QUESTION = "sensitive_question"
+    CONTEXT_ANONYMOUS_QUESTION = "anonymous_question"
     CONTEXT_ASKED_QUESTION = "asked_question"
     CONTEXT_QUESTION_TO_ANSWER = "question_to_answer"
+    CONTEXT_SENSITIVE_QUESTION_TO_ANSWER = "sensitive_question_to_answer"
     CONTEXT_MESSAGE_TO_REPORT = "message_to_report"
     CONTEXT_REPORTING_IS_QUESTION = "reporting_is_question"
     CONTEXT_REPORTING_REASON = "reporting_reason"
@@ -64,6 +66,10 @@ class AskForHelpHandler(WenetEventHandler):
     INTENT_ASK_TO_ANYONE = "ask_to_anyone"
     INTENT_ASK_TO_NEARBY = "ask_to_nearby"
     INTENT_ASK_TO_ANYWHERE = "ask_to_anywhere"
+    INTENT_SENSITIVE_QUESTION = "sensitive"
+    INTENT_NOT_SENSITIVE_QUESTION = "not_sensitive"
+    INTENT_ANONYMOUS_QUESTION = "anonymous"
+    INTENT_NOT_ANONYMOUS_QUESTION = "not_anonymous"
     INTENT_ANSWER_QUESTION = "answer_question"
     INTENT_ANSWER_REMIND_LATER = "answer_remind_later"
     INTENT_ANSWER_NOT = "answer_not"
@@ -81,6 +87,8 @@ class AskForHelpHandler(WenetEventHandler):
     STATE_QUESTION_2 = "question_2"
     STATE_QUESTION_3 = "question_3"
     STATE_QUESTION_4 = "question_4"
+    STATE_QUESTION_4_1 = "question_4_1"
+    STATE_QUESTION_5 = "question_5"
     STATE_ANSWERING = "answer_2"
     # transaction labels
     LABEL_ANSWER_TRANSACTION = "answerTransaction"
@@ -140,15 +148,39 @@ class AskForHelpHandler(WenetEventHandler):
                 .with_rule(static_context=(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_3))
         )
         self.intent_manager.with_fulfiller(
-            IntentFulfillerV3(self.INTENT_ASK_TO_NEARBY, self.action_question_5).with_rule(
-                intent=self.INTENT_ASK_TO_NEARBY,
+            IntentFulfillerV3(self.INTENT_SENSITIVE_QUESTION, self.action_question_4_1).with_rule(
+                intent=self.INTENT_SENSITIVE_QUESTION,
                 static_context=(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_4)
             )
         )
         self.intent_manager.with_fulfiller(
-            IntentFulfillerV3(self.INTENT_ASK_TO_ANYWHERE, self.action_question_5).with_rule(
-                intent=self.INTENT_ASK_TO_ANYWHERE,
+            IntentFulfillerV3(self.INTENT_NOT_SENSITIVE_QUESTION, self.action_question_5).with_rule(
+                intent=self.INTENT_NOT_SENSITIVE_QUESTION,
                 static_context=(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_4)
+            )
+        )
+        self.intent_manager.with_fulfiller(
+            IntentFulfillerV3(self.INTENT_ANONYMOUS_QUESTION, self.action_question_5).with_rule(
+                intent=self.INTENT_ANONYMOUS_QUESTION,
+                static_context=(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_4_1)
+            )
+        )
+        self.intent_manager.with_fulfiller(
+            IntentFulfillerV3(self.INTENT_NOT_ANONYMOUS_QUESTION, self.action_question_5).with_rule(
+                intent=self.INTENT_NOT_ANONYMOUS_QUESTION,
+                static_context=(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_4_1)
+            )
+        )
+        self.intent_manager.with_fulfiller(
+            IntentFulfillerV3(self.INTENT_ASK_TO_NEARBY, self.action_question_6).with_rule(
+                intent=self.INTENT_ASK_TO_NEARBY,
+                static_context=(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_5)
+            )
+        )
+        self.intent_manager.with_fulfiller(
+            IntentFulfillerV3(self.INTENT_ASK_TO_ANYWHERE, self.action_question_6).with_rule(
+                intent=self.INTENT_ASK_TO_ANYWHERE,
+                static_context=(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_5)
             )
         )
         self.intent_manager.with_fulfiller(
@@ -211,10 +243,11 @@ class AskForHelpHandler(WenetEventHandler):
         return response
 
     def _clear_context(self, context: ConversationContext) -> ConversationContext:
-        context_to_remove = [self.CONTEXT_CURRENT_STATE, self.CONTEXT_ASKED_QUESTION, self.CONTEXT_DESIRED_ANSWERER,
-                             self.CONTEXT_QUESTION_TO_ANSWER, self.CONTEXT_MESSAGE_TO_REPORT,
-                             self.CONTEXT_REPORTING_IS_QUESTION, self.CONTEXT_REPORTING_REASON,
-                             self.CONTEXT_ORIGINAL_QUESTION_REPORTING, self.CONTEXT_DESIRED_ANSWERER_REASON]
+        context_to_remove = [
+            self.CONTEXT_CURRENT_STATE, self.CONTEXT_ASKED_QUESTION, self.CONTEXT_DESIRED_ANSWERER,
+            self.CONTEXT_QUESTION_TO_ANSWER, self.CONTEXT_SENSITIVE_QUESTION_TO_ANSWER, self.CONTEXT_MESSAGE_TO_REPORT,
+            self.CONTEXT_REPORTING_IS_QUESTION, self.CONTEXT_REPORTING_REASON, self.CONTEXT_ORIGINAL_QUESTION_REPORTING,
+            self.CONTEXT_DESIRED_ANSWERER_REASON, self.CONTEXT_SENSITIVE_QUESTION, self.CONTEXT_ANONYMOUS_QUESTION]
         for context_key in context_to_remove:
             context.delete_static_state(context_key)
         return context
@@ -223,7 +256,8 @@ class AskForHelpHandler(WenetEventHandler):
         """
         Returns True if the user is in another action (e.g. inside the /question flow), False otherwise
         """
-        statuses = [self.STATE_ANSWERING, self.STATE_QUESTION_1, self.STATE_QUESTION_2, self.STATE_QUESTION_3, self.STATE_QUESTION_4]
+        statuses = [self.STATE_ANSWERING, self.STATE_QUESTION_1, self.STATE_QUESTION_2, self.STATE_QUESTION_3,
+                    self.STATE_QUESTION_4, self.STATE_QUESTION_4_1, self.STATE_QUESTION_5]
         current_status = context.get_static_state(self.CONTEXT_CURRENT_STATE, "")
         return current_status in statuses
 
@@ -297,19 +331,26 @@ class AskForHelpHandler(WenetEventHandler):
         response = TelegramTextualResponse(f"{title}_{self.parse_text_with_markdown(message.text)}_")
         return NotificationEvent(user_account.social_details, [response], user_account.context)
 
-    def handle_nearby_question(self, message: QuestionToAnswerMessage, user_object: WeNetUserProfile, questioning_user: WeNetUserProfile) -> TelegramRapidAnswerResponse:
-        # Translate the message that someone near has a question and insert the details of the question
-        message_string = self._translator.get_translation_instance(user_object.locale) \
-            .with_text("answer_message_nearby") \
-            .with_substitution("question", self.parse_text_with_markdown(message.question)) \
-            .with_substitution("user", questioning_user.name.first if questioning_user.name.first else "Anonymous") \
+    def handle_nearby_question(self, message: QuestionToAnswerMessage, user_object: WeNetUserProfile,
+                               questioning_user: WeNetUserProfile, sensitive: bool, anonymous: bool) -> TelegramRapidAnswerResponse:
+        # Translate the message that someone near has a question and insert the details of the question, treat differently sensitive questions
+        message_string = self._translator.get_translation_instance(user_object.locale)
+        if sensitive:
+            message_string = message_string.with_text("answer_sensitive_message_nearby")
+        else:
+            message_string = message_string.with_text("answer_message_nearby")
+
+        message_string = message_string.with_substitution("question", self.parse_text_with_markdown(message.question)) \
+            .with_substitution("user", questioning_user.name.first if questioning_user.name.first and not anonymous else "Anonymous") \
             .translate()
+
         # we create ids of all buttons, to know which buttons invalidate when one of them is clicked
         button_ids = [str(uuid.uuid4()) for _ in range(3)]
         button_data = {
             "task_id": message.task_id,
             "question": message.question,
-            "username": questioning_user.name.first if questioning_user.name.first else "Anonymous",
+            "sensitive": sensitive,
+            "username": questioning_user.name.first if questioning_user.name.first and not anonymous else "Anonymous",
             "related_buttons": button_ids,
         }
         response = TelegramRapidAnswerResponse(TextualResponse(message_string), row_displacement=[1, 1, 1])
@@ -321,19 +362,26 @@ class AskForHelpHandler(WenetEventHandler):
         response.with_textual_option(self._translator.get_translation_instance(user_object.locale).with_text("answer_report_button").translate(), self.INTENT_BUTTON_WITH_PAYLOAD.format(button_ids[2]))
         return response
 
-    def handle_question(self, message: QuestionToAnswerMessage, user_object: WeNetUserProfile, questioning_user: WeNetUserProfile) -> TelegramRapidAnswerResponse:
-        # Translate the message that someone in the community has a question and insert the details of the question
-        message_string = self._translator.get_translation_instance(user_object.locale) \
-            .with_text("answer_message_0") \
-            .with_substitution("question", self.parse_text_with_markdown(message.question)) \
-            .with_substitution("user", questioning_user.name.first if questioning_user.name.first else "Anonymous") \
+    def handle_question(self, message: QuestionToAnswerMessage, user_object: WeNetUserProfile,
+                        questioning_user: WeNetUserProfile, sensitive: bool, anonymous: bool) -> TelegramRapidAnswerResponse:
+        # Translate the message that someone in the community has a question and insert the details of the question, treat differently sensitive questions
+        message_string = self._translator.get_translation_instance(user_object.locale)
+        if sensitive:
+            message_string = message_string.with_text("answer_sensitive_message_0")
+        else:
+            message_string = message_string.with_text("answer_message_0")
+
+        message_string = message_string.with_substitution("question", self.parse_text_with_markdown(message.question)) \
+            .with_substitution("user", questioning_user.name.first if questioning_user.name.first and not anonymous else "Anonymous") \
             .translate()
+
         # we create ids of all buttons, to know which buttons invalidate when one of them is clicked
         button_ids = [str(uuid.uuid4()) for _ in range(4)]
         button_data = {
             "task_id": message.task_id,
             "question": message.question,
-            "username": questioning_user.name.first if questioning_user.name.first else "Anonymous",
+            "sensitive": sensitive,
+            "username": questioning_user.name.first if questioning_user.name.first and not anonymous else "Anonymous",
             "related_buttons": button_ids,
         }
         response = TelegramRapidAnswerResponse(TextualResponse(message_string), row_displacement=[1, 1, 1, 1])
@@ -357,6 +405,9 @@ class AskForHelpHandler(WenetEventHandler):
             .with_substitution("answer", self.parse_text_with_markdown(answer_text)) \
             .with_substitution("username", answerer_user.name.first if answerer_user.name.first else "Anonymous") \
             .translate()
+
+        # TODO handle anonymous answerer
+
         answer = TelegramRapidAnswerResponse(TextualResponse(message_string), row_displacement=[1, 1, 1])
         button_report_text = self._translator.get_translation_instance(user_object.locale).with_text("answer_report_button").translate()
         button_more_answers_text = self._translator.get_translation_instance(user_object.locale).with_text("more_answers_button").translate()
@@ -403,10 +454,15 @@ class AskForHelpHandler(WenetEventHandler):
                 # handle a new question to answer checking if the question is for nearby people
                 questioning_user = service_api.get_user_profile(str(message.user_id))
                 question_task = service_api.get_task(message.task_id)
+
+                #  Extract sensitive and anonymous attributes
+                sensitive = question_task.attributes.get("sensitive")
+                anonymous = question_task.attributes.get("anonymous")
+
                 if question_task.attributes.get("positionOfAnswerer") == self.INTENT_ASK_TO_NEARBY:
-                    response = self.handle_nearby_question(message, user_object, questioning_user)
+                    response = self.handle_nearby_question(message, user_object, questioning_user, sensitive, anonymous)
                 else:
-                    response = self.handle_question(message, user_object, questioning_user)
+                    response = self.handle_question(message, user_object, questioning_user, sensitive, anonymous)
                 return NotificationEvent(user_account.social_details, [response], context)
             elif isinstance(message, AnsweredQuestionMessage):
                 # handle an answer to a question
@@ -566,7 +622,7 @@ class AskForHelpHandler(WenetEventHandler):
 
     def action_question_4(self, incoming_event: IncomingSocialEvent, _: str) -> OutgoingEvent:
         """
-        Save the why this type of desired answerer, and ask where is the person that should answer the question
+        Save the why this type of desired answerer, and ask whether the question is sensitive or not
         """
         user_locale = self._get_user_locale_from_incoming_event(incoming_event)
         response = OutgoingEvent(social_details=incoming_event.social_details)
@@ -575,12 +631,12 @@ class AskForHelpHandler(WenetEventHandler):
             context = incoming_event.context
             context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_4)
             context.with_static_state(self.CONTEXT_DESIRED_ANSWERER_REASON, desired_answerer_reason)
-            message = self._translator.get_translation_instance(user_locale).with_text("specify_answerer_location").translate()
-            button_1_text = self._translator.get_translation_instance(user_locale).with_text("location_answer_1").translate()
-            button_2_text = self._translator.get_translation_instance(user_locale).with_text("location_answer_2").translate()
+            message = self._translator.get_translation_instance(user_locale).with_text("sensitive_question").translate()
+            button_1_text = self._translator.get_translation_instance(user_locale).with_text("not_sensitive").translate()
+            button_2_text = self._translator.get_translation_instance(user_locale).with_text("sensitive").translate()
             response_with_buttons = TelegramRapidAnswerResponse(TextualResponse(message), row_displacement=[1, 1])
-            response_with_buttons.with_textual_option(button_1_text, self.INTENT_ASK_TO_NEARBY)
-            response_with_buttons.with_textual_option(button_2_text, self.INTENT_ASK_TO_ANYWHERE)
+            response_with_buttons.with_textual_option(button_1_text, self.INTENT_NOT_SENSITIVE_QUESTION)
+            response_with_buttons.with_textual_option(button_2_text, self.INTENT_SENSITIVE_QUESTION)
             response.with_message(response_with_buttons)
             response.with_context(context)
         else:
@@ -588,7 +644,48 @@ class AskForHelpHandler(WenetEventHandler):
             response.with_message(TextualResponse(error_message))
         return response
 
+    def action_question_4_1(self, incoming_event: IncomingSocialEvent, intent: str) -> OutgoingEvent:
+        """
+        Save whether the question is sensitive or not, and ask whether to ask the question anonymously or not
+        """
+        user_locale = self._get_user_locale_from_incoming_event(incoming_event)
+        response = OutgoingEvent(social_details=incoming_event.social_details)
+        context = incoming_event.context
+        context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_4_1)
+        context.with_static_state(self.CONTEXT_SENSITIVE_QUESTION, intent)
+        message = self._translator.get_translation_instance(user_locale).with_text("anonymous_question").translate()
+        button_1_text = self._translator.get_translation_instance(user_locale).with_text("anonymous").translate()
+        button_2_text = self._translator.get_translation_instance(user_locale).with_text("not_anonymous").translate()
+        response_with_buttons = TelegramRapidAnswerResponse(TextualResponse(message), row_displacement=[1, 1])
+        response_with_buttons.with_textual_option(button_1_text, self.INTENT_ANONYMOUS_QUESTION)
+        response_with_buttons.with_textual_option(button_2_text, self.INTENT_NOT_ANONYMOUS_QUESTION)
+        response.with_message(response_with_buttons)
+        response.with_context(context)
+        return response
+
     def action_question_5(self, incoming_event: IncomingSocialEvent, intent: str) -> OutgoingEvent:
+        """
+        Save whether the question is not sensitive or whether to ask the question anonymously or not, and ask where should be the person that should answer the question
+        """
+        user_locale = self._get_user_locale_from_incoming_event(incoming_event)
+        response = OutgoingEvent(social_details=incoming_event.social_details)
+        context = incoming_event.context
+        context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_5)
+        if intent in [self.INTENT_ANONYMOUS_QUESTION,  self.INTENT_NOT_ANONYMOUS_QUESTION]:
+            context.with_static_state(self.CONTEXT_ANONYMOUS_QUESTION, intent)
+        else:
+            context.with_static_state(self.CONTEXT_SENSITIVE_QUESTION, intent)
+        message = self._translator.get_translation_instance(user_locale).with_text("specify_answerer_location").translate()
+        button_1_text = self._translator.get_translation_instance(user_locale).with_text("location_answer_1").translate()
+        button_2_text = self._translator.get_translation_instance(user_locale).with_text("location_answer_2").translate()
+        response_with_buttons = TelegramRapidAnswerResponse(TextualResponse(message), row_displacement=[1, 1])
+        response_with_buttons.with_textual_option(button_1_text, self.INTENT_ASK_TO_NEARBY)
+        response_with_buttons.with_textual_option(button_2_text, self.INTENT_ASK_TO_ANYWHERE)
+        response.with_message(response_with_buttons)
+        response.with_context(context)
+        return response
+
+    def action_question_6(self, incoming_event: IncomingSocialEvent, intent: str) -> OutgoingEvent:
         """
         Conclude the /question flow, with a final message
         """
@@ -599,15 +696,23 @@ class AskForHelpHandler(WenetEventHandler):
         user_locale = self._get_user_locale_from_incoming_event(incoming_event)
         response = OutgoingEvent(social_details=incoming_event.social_details)
         context = incoming_event.context
-        if not context.has_static_state(self.CONTEXT_ASKED_QUESTION) or not context.has_static_state(self.CONTEXT_DESIRED_ANSWERER) or not context.has_static_state(self.CONTEXT_DESIRED_ANSWERER_REASON):
-            raise Exception(f"Expected {self.CONTEXT_ASKED_QUESTION}, {self.CONTEXT_DESIRED_ANSWERER} and {self.CONTEXT_DESIRED_ANSWERER_REASON} in the context")
+        if not context.has_static_state(self.CONTEXT_ASKED_QUESTION) \
+                or not context.has_static_state(self.CONTEXT_DESIRED_ANSWERER) \
+                or not context.has_static_state(self.CONTEXT_DESIRED_ANSWERER_REASON) \
+                or not context.has_static_state(self.CONTEXT_SENSITIVE_QUESTION):
+            raise Exception(f"Expected {self.CONTEXT_ASKED_QUESTION}, {self.CONTEXT_DESIRED_ANSWERER}, "
+                            f"{self.CONTEXT_DESIRED_ANSWERER_REASON} and {self.CONTEXT_SENSITIVE_QUESTION} in the context")
         wenet_id = context.get_static_state(self.CONTEXT_WENET_USER_ID)
         question = context.get_static_state(self.CONTEXT_ASKED_QUESTION)
         desired_answerer = context.get_static_state(self.CONTEXT_DESIRED_ANSWERER)
         answerer_details = context.get_static_state(self.CONTEXT_DESIRED_ANSWERER_REASON)
+        sensitive = context.get_static_state(self.CONTEXT_SENSITIVE_QUESTION)
+        anonymous = context.get_static_state(self.CONTEXT_ANONYMOUS_QUESTION, self.INTENT_NOT_ANONYMOUS_QUESTION)
         attributes = {
             "kindOfAnswerer": desired_answerer,
             "answeredDetails": answerer_details,
+            "sensitive": True if sensitive == self.INTENT_SENSITIVE_QUESTION else False,
+            "anonymous": True if anonymous == self.INTENT_ANONYMOUS_QUESTION else False,
             "positionOfAnswerer": intent,
         }
         question_task = Task(
@@ -627,7 +732,7 @@ class AskForHelpHandler(WenetEventHandler):
         try:
             service_api.create_task(question_task)
             logger.debug(f"User [{wenet_id}] asked a question. Task created successfully")
-            message = self._translator.get_translation_instance(user_locale).with_text("question_5").translate()
+            message = self._translator.get_translation_instance(user_locale).with_text("question_final").translate()
             response.with_message(TextualResponse(message))
         except TaskCreationError as e:
             logger.error(f"The service API responded with code {e.http_status} and message {json.dumps(e.json_response)}")
@@ -637,6 +742,8 @@ class AskForHelpHandler(WenetEventHandler):
             context.delete_static_state(self.CONTEXT_ASKED_QUESTION)
             context.delete_static_state(self.CONTEXT_DESIRED_ANSWERER)
             context.delete_static_state(self.CONTEXT_DESIRED_ANSWERER_REASON)
+            context.delete_static_state(self.CONTEXT_SENSITIVE_QUESTION)
+            context.delete_static_state(self.CONTEXT_ANONYMOUS_QUESTION)
             context.delete_static_state(self.CONTEXT_CURRENT_STATE)
             response.with_context(context)
         return response
@@ -658,15 +765,18 @@ class AskForHelpHandler(WenetEventHandler):
         """
         user_locale = self._get_user_locale_from_incoming_event(incoming_event)
         context = incoming_event.context
-        user_id = context.get_static_state(self.CONTEXT_WENET_USER_ID)
-        show_conduct_message = True
-        if user_id:
-            is_first_answer = self.is_first_answer(user_id)
-            show_conduct_message = is_first_answer or random.randint(1, 10) <= 2
+        # user_id = context.get_static_state(self.CONTEXT_WENET_USER_ID)
+        # show_conduct_message = True
+        # if user_id:
+        #     is_first_answer = self.is_first_answer(user_id)
+        #     show_conduct_message = is_first_answer or random.randint(1, 10) <= 2
         context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_ANSWERING)
         context.with_static_state(self.CONTEXT_QUESTION_TO_ANSWER, button_payload.payload["task_id"])
-        if show_conduct_message:
-            message = self._translator.get_translation_instance(user_locale).with_text("question_0").translate()
+        context.with_static_state(self.CONTEXT_SENSITIVE_QUESTION_TO_ANSWER, button_payload.payload.get("sensitive"))
+        # if show_conduct_message:
+        #     message = self._translator.get_translation_instance(user_locale).with_text("question_0").translate()
+        if button_payload.payload.get("sensitive"):
+            message = self._translator.get_translation_instance(user_locale).with_text("answer_sensitive_question").translate()
         else:
             message = self._translator.get_translation_instance(user_locale).with_text("answer_question").translate()
         response = OutgoingEvent(social_details=incoming_event.social_details)
@@ -703,7 +813,7 @@ class AskForHelpHandler(WenetEventHandler):
 
     def action_answer_question_2(self, incoming_event: IncomingSocialEvent, _: str) -> OutgoingEvent:
         """
-        QuestionToAnswerMessage flow, collect the user's answer and thank her
+        QuestionToAnswerMessage flow, collect the user's answer and thank her. If it is sensitive ask if ahould be anonymous the answer
         """
         if incoming_event.context is not None:
             service_api = self._get_service_api_interface_connector_from_context(incoming_event.context)
@@ -712,6 +822,9 @@ class AskForHelpHandler(WenetEventHandler):
 
         user_locale = self._get_user_locale_from_incoming_event(incoming_event)
         context = incoming_event.context
+
+        # TODO handle anonymous answerer flow
+
         if not context.has_static_state(self.CONTEXT_QUESTION_TO_ANSWER):
             error_message = "Illegal state, expected the question ID in the context, but it does not exist"
             logger.error(error_message)
@@ -722,9 +835,7 @@ class AskForHelpHandler(WenetEventHandler):
             answer = incoming_event.incoming_message.text
             actioneer_id = context.get_static_state(self.CONTEXT_WENET_USER_ID)
             try:
-                transaction = TaskTransaction(None, question_id, self.LABEL_ANSWER_TRANSACTION,
-                                              int(datetime.now().timestamp()), int(datetime.now().timestamp()),
-                                              actioneer_id, {"answer": answer}, [])
+                transaction = TaskTransaction(None, question_id, self.LABEL_ANSWER_TRANSACTION, int(datetime.now().timestamp()), int(datetime.now().timestamp()), actioneer_id, {"answer": answer}, [])
                 service_api.create_task_transaction(transaction)
                 logger.info("Sent task transaction: %s" % str(transaction.to_repr()))
                 message = self._translator.get_translation_instance(user_locale).with_text("answered_message").translate()
@@ -754,9 +865,7 @@ class AskForHelpHandler(WenetEventHandler):
         question_id = button_payload.payload["task_id"]
         actioneer_id = context.get_static_state(self.CONTEXT_WENET_USER_ID)
         try:
-            transaction = TaskTransaction(None, question_id, self.LABEL_NOT_ANSWER_TRANSACTION,
-                                          int(datetime.now().timestamp()), int(datetime.now().timestamp()),
-                                          actioneer_id, {}, [])
+            transaction = TaskTransaction(None, question_id, self.LABEL_NOT_ANSWER_TRANSACTION, int(datetime.now().timestamp()), int(datetime.now().timestamp()), actioneer_id, {}, [])
             service_api.create_task_transaction(transaction)
             message = self._translator.get_translation_instance(user_locale).with_text("not_answer_response").translate()
             response.with_message(TextualResponse(message))
@@ -776,12 +885,18 @@ class AskForHelpHandler(WenetEventHandler):
         response.with_message(TextualResponse(message))
         pending_answers = context.get_static_state(self.CONTEXT_PENDING_ANSWERS, {})
         question_id = button_payload.payload["task_id"]
-        # recreating the message
-        message_string = self._translator.get_translation_instance(user_locale) \
-            .with_text("answer_message_0") \
-            .with_substitution("question", self.parse_text_with_markdown(button_payload.payload["question"])) \
+
+        # Recreating the message that someone in the community has a question and insert the details of the question, treat differently sensitive questions
+        message_string = self._translator.get_translation_instance(user_locale)
+        if button_payload.payload.get("sensitive"):
+            message_string = message_string.with_text("answer_sensitive_message_0")
+        else:
+            message_string = message_string.with_text("answer_message_0")
+
+        message_string = message_string.with_substitution("question", self.parse_text_with_markdown(button_payload.payload["question"])) \
             .with_substitution("user", button_payload.payload["username"]) \
             .translate()
+
         button_ids = [str(uuid.uuid4()) for _ in range(4)]
         button_data = {
             "task_id": question_id,
@@ -792,21 +907,14 @@ class AskForHelpHandler(WenetEventHandler):
         response_to_store = TelegramRapidAnswerResponse(TextualResponse(message_string), row_displacement=[1, 1, 1, 1])
 
         self.cache.cache(ButtonPayload(button_data, self.INTENT_ANSWER_QUESTION).to_repr(), key=button_ids[0])
-        response_to_store.with_textual_option(self._translator.get_translation_instance(user_locale).with_text(
-            "answer_question_button").translate(), self.INTENT_BUTTON_WITH_PAYLOAD.format(button_ids[0]))
+        response_to_store.with_textual_option(self._translator.get_translation_instance(user_locale).with_text("answer_question_button").translate(), self.INTENT_BUTTON_WITH_PAYLOAD.format(button_ids[0]))
         self.cache.cache(ButtonPayload(button_data, self.INTENT_ANSWER_REMIND_LATER).to_repr(), key=button_ids[1])
-        response_to_store.with_textual_option(self._translator.get_translation_instance(user_locale).with_text(
-            "answer_remind_later_button").translate(), self.INTENT_BUTTON_WITH_PAYLOAD.format(button_ids[1]))
+        response_to_store.with_textual_option(self._translator.get_translation_instance(user_locale).with_text("answer_remind_later_button").translate(), self.INTENT_BUTTON_WITH_PAYLOAD.format(button_ids[1]))
         self.cache.cache(ButtonPayload(button_data, self.INTENT_ANSWER_NOT).to_repr(), key=button_ids[2])
-        response_to_store.with_textual_option(
-            self._translator.get_translation_instance(user_locale).with_text("answer_not_button").translate(),
-            self.INTENT_BUTTON_WITH_PAYLOAD.format(button_ids[2]))
+        response_to_store.with_textual_option(self._translator.get_translation_instance(user_locale).with_text("answer_not_button").translate(), self.INTENT_BUTTON_WITH_PAYLOAD.format(button_ids[2]))
         self.cache.cache(ButtonPayload(button_data, self.INTENT_QUESTION_REPORT).to_repr(), key=button_ids[3])
-        response_to_store.with_textual_option(
-            self._translator.get_translation_instance(user_locale).with_text("answer_report_button").translate(),
-            self.INTENT_BUTTON_WITH_PAYLOAD.format(button_ids[3]))
-        pending_answer = PendingQuestionToAnswer(question_id, response_to_store, incoming_event.social_details,
-                                                 sent=datetime.now())
+        response_to_store.with_textual_option(self._translator.get_translation_instance(user_locale).with_text("answer_report_button").translate(), self.INTENT_BUTTON_WITH_PAYLOAD.format(button_ids[3]))
+        pending_answer = PendingQuestionToAnswer(question_id, response_to_store, incoming_event.social_details, sent=datetime.now())
         pending_answers[question_id] = pending_answer.to_repr()
         context.with_static_state(self.CONTEXT_PENDING_ANSWERS, pending_answers)
         response.with_context(context)
@@ -855,9 +963,7 @@ class AskForHelpHandler(WenetEventHandler):
             attributes.update({"transactionId": transaction_id})
         actioneer_id = incoming_event.context.get_static_state(self.CONTEXT_WENET_USER_ID)
         try:
-            transaction = TaskTransaction(None, task_id, transaction_label,
-                                          int(datetime.now().timestamp()), int(datetime.now().timestamp()),
-                                          actioneer_id, attributes, [])
+            transaction = TaskTransaction(None, task_id, transaction_label, int(datetime.now().timestamp()), int(datetime.now().timestamp()), actioneer_id, attributes, [])
             service_api.create_task_transaction(transaction)
             logger.info("Sent task transaction: %s" % str(transaction.to_repr()))
             message = self._translator.get_translation_instance(user_locale).with_text(
@@ -881,9 +987,7 @@ class AskForHelpHandler(WenetEventHandler):
         task_id = button_payload.payload["task_id"]
         actioneer_id = context.get_static_state(self.CONTEXT_WENET_USER_ID)
         try:
-            transaction = TaskTransaction(None, task_id, self.LABEL_MORE_ANSWER_TRANSACTION,
-                                          int(datetime.now().timestamp()), int(datetime.now().timestamp()),
-                                          actioneer_id, {}, [])
+            transaction = TaskTransaction(None, task_id, self.LABEL_MORE_ANSWER_TRANSACTION, int(datetime.now().timestamp()), int(datetime.now().timestamp()), actioneer_id, {}, [])
             service_api.create_task_transaction(transaction)
             logger.info("Sent task transaction: %s" % str(transaction.to_repr()))
             message = self._translator.get_translation_instance(user_locale).with_text("ask_more_answers_text").translate()
@@ -909,9 +1013,7 @@ class AskForHelpHandler(WenetEventHandler):
         }
         actioneer_id = incoming_event.context.get_static_state(self.CONTEXT_WENET_USER_ID)
         try:
-            transaction = TaskTransaction(None, task_id, self.LABEL_BEST_ANSWER_TRANSACTION,
-                                          int(datetime.now().timestamp()), int(datetime.now().timestamp()),
-                                          actioneer_id, attributes, [])
+            transaction = TaskTransaction(None, task_id, self.LABEL_BEST_ANSWER_TRANSACTION, int(datetime.now().timestamp()), int(datetime.now().timestamp()), actioneer_id, attributes, [])
             service_api.create_task_transaction(transaction)
             logger.info("Sent task transaction: %s" % str(transaction.to_repr()))
             message = self._translator.get_translation_instance(user_locale).with_text(
@@ -936,6 +1038,9 @@ class AskForHelpHandler(WenetEventHandler):
         tasks = [t for t in service_api.get_all_tasks_of_application(self.app_id)
                  if t.requester_id != user_id and user_id not in set(
                 [transaction.actioneer_id for transaction in t.transactions if transaction.label == "answerTransaction"])]
+
+        # TODO handle sensitive and anonymous questions
+
         if not tasks:
             response.with_message(TextualResponse(
                 self._translator.get_translation_instance(user_locale).with_text("answers_no_tasks").translate()))
@@ -967,4 +1072,3 @@ class AskForHelpHandler(WenetEventHandler):
         text = self._translator.get_translation_instance(user_locale).with_text("not_implemented").translate()
         response.with_message(TextualResponse(text))
         return response
-