+++ changelog.md
@@ -3,7 +3,8 @@
 # Versions
 
 ## next
- - Integrated position data in the question and answer flow
+- Integrated position data in the question and answer flow
+- Integrated sensitive and anonymous information data in the question and answer flow
 
 ## 1.0.10
 - Changed the `project` field in the messages logged into the Wenet platform. Now the `appId` is used, instead of the `PROJECT_NAME` environment variable